# API契約.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "簡易POS API契約案",
    "version": "0.37",
    "description": "新要件に基づく設計付属物。未実装・未実行。URL／型／必須項目の正本。業務状態・トランザクションは設計仕様書。金額・版は十進文字列。未定義入力を拒否。成功・失敗ともno-store。 認証・権限はAPI別securityと設計8.2に従う。Origin・X-POS-Requestの必須検査は全更新API。GETでは両ヘッダーの欠落を拒否理由にしない。"
  },
  "servers": [
    {
      "url": "/",
      "description": "Next.jsと同一オリジン。FastAPI直通は公開クライアント契約外。"
    }
  ],
  "paths": {
...
    "/api/register/start": {
      "post": {
        "operationId": "registerStart",
        "summary": "初回Cookie発行",
        "security": [
          {
            "AuthCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-POS-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "許可されたPOSオリジンと完全一致。サーバー中継は元Originを検証して転送。"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "保護Cookie。認証は__Host-pos_session、復帰は__Host-pos_resume。複数行を統合せず転送。値は本文へ含めない。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Register"
                }
              }
            }
          },
          "401": {
            "description": "未認証／誤認証",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "権限・送信元・Cookie対応不正",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "対象不存在",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "状態／版／期限／保守停止等の競合",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "入力不正",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "DB等の照合不能",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "予期しないサーバー障害",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "設計仕様書5章・8章の権限／状態／再送条件を適用。 UNSTARTEDだけ発行。開始済みの再送は409でstatus照会へ。識別子を再発行しない。"
      }
    },
...
    "/api/register/confirm": {
      "post": {
        "operationId": "registerConfirm",
        "summary": "Cookie受取確認",
        "security": [
          {
            "AuthCookie": [],
            "ResumeCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-POS-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "許可されたPOSオリジンと完全一致。サーバー中継は元Originを検証して転送。"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Register"
                }
              }
            }
          },
          "401": {
            "description": "未認証／誤認証",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "権限・送信元・Cookie対応不正",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "対象不存在",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "状態／版／期限／保守停止等の競合",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "入力不正",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "DB等の照合不能",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "予期しないサーバー障害",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "設計仕様書5章・8章の権限／状態／再送条件を適用。"
      }
    },
    "/api/carts": {
      "post": {
        "operationId": "createCart",
        "summary": "初回カート生成",
        "security": [
          {
            "AuthCookie": [],
            "ResumeCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-POS-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "許可されたPOSオリジンと完全一致。サーバー中継は元Originを検証して転送。"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResult"
                }
              }
            }
          },
          "401": {
            "description": "未認証／誤認証",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "権限・送信元・Cookie対応不正",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "対象不存在",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "状態／版／期限／保守停止等の競合",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "入力不正",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "DB等の照合不能",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "予期しないサーバー障害",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "成功・失敗ともキャッシュしない"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "設計仕様書5章・8章の権限／状態／再送条件を適用。"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pos_session"
      },
      "ResumeCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pos_resume"
      }
    },
    "schemas": {
      "Code": {
        "type": "string",
        "minLength": 1,
        "maxLength": 32,
        "pattern": "^[A-Za-z0-9_-]+(?![\\s\\S])"
      },
...
      "Register": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "start_state": {
            "type": "string",
            "enum": [
              "UNSTARTED",
              "COOKIE_PENDING",
              "READY"
            ]
          },
          "cookie_status": {
            "type": "string",
            "enum": [
              "ABSENT",
              "MATCH",
              "INVALID"
            ]
          },
          "available": {
            "type": "boolean"
          },
          "maintenance_hold": {
            "type": "boolean"
          }
        },
        "required": [
          "start_state",
          "cookie_status",
          "available",
          "maintenance_hold"
        ]
      },
...
      "Purchase": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "cart_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "staff_id": {
            "$ref": "#/components/schemas/Code"
          },
          "member_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Code"
              },
              {
                "type": "null"
              }
            ]
          },
          "purchased_at": {
            "$ref": "#/components/schemas/Time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseLine"
            },
            "minItems": 1
          },
          "subtotal": {
            "$ref": "#/components/schemas/Money"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "taxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tax"
            },
            "minItems": 1
          }
        },
        "required": [
          "cart_id",
          "staff_id",
          "member_id",
          "purchased_at",
          "lines",
          "subtotal",
          "total",
          "taxes"
        ]
      },
      "CartResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "cart": {
            "$ref": "#/components/schemas/Cart"
          },
          "purchase_status": {
            "type": "string",
            "enum": [
              "SAVED",
              "UNSAVED",
              "UNKNOWN",
              "NOT_REQUESTED"
            ]
          },
          "purchase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Purchase"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "cart",
          "purchase_status",
          "purchase"
        ],
        "allOf": [
          {
            "if": {
              "properties": {
                "purchase_status": {
                  "const": "SAVED"
                }
              }
            },
            "then": {
              "properties": {
                "purchase": {
                  "$ref": "#/components/schemas/Purchase"
                }
              }
            },
            "else": {
              "properties": {
                "purchase": {
                  "type": "null"
                }
              }
            }
          }
        ]
      },
      "OperationResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "cart": {
            "$ref": "#/components/schemas/Cart"
          },
          "purchase_status": {
            "type": "string",
            "enum": [
              "SAVED",
              "UNSAVED",
              "UNKNOWN",
              "NOT_REQUESTED"
            ]
          },
          "purchase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Purchase"
              },
              {
                "type": "null"
              }
            ]
          },
          "operation_id": {
            "$ref": "#/components/schemas/UUID"
          },
          "operation_status": {
            "type": "string",
            "enum": [
              "PREPARED",
              "APPLIED",
              "REJECTED",
              "NOT_FOUND"
            ]
          },
          "applied_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Version"
              },
              {
                "type": "null"
              }
            ]
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "拒否・失敗理由、またはresolve-purchaseの会員PENDING分岐成功を示すPURCHASE_FENCED_MEMBER_PENDING。再送・操作GETでも当該操作の記録値を返す。"
          },
          "new_cart_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UUID"
              },
              {
                "type": "null"
              }
            ],
            "description": "操作GETでは必須。NEXT反映済みの場合だけCART_OPERATION.next_cart_idの記録済みID、それ以外はnull。他のOperationResult応答では省略可。現在の継続対象はresumeで確認する。"
          }
        },
        "required": [
          "cart",
          "purchase_status",
          "purchase",
          "operation_id",
          "operation_status",
          "applied_version",
          "code"
        ],
        "allOf": [
          {
            "if": {
              "properties": {
                "purchase_status": {
                  "const": "SAVED"
                }
              }
            },
            "then": {
              "properties": {
                "purchase": {
                  "$ref": "#/components/schemas/Purchase"
                }
              }
            },
            "else": {
              "properties": {
                "purchase": {
                  "type": "null"
                }
              }
            }
          }
        ],
        "examples": [
          {
            "cart": {
              "cart_id": "11111111-1111-4111-8111-111111111111",
              "version": "2",
              "state": "EDITING",
              "staff_id": "staff01",
              "member_state": "UNSPECIFIED",
              "member_id": null,
              "pending_member_id": null,
              "lines": [
                {
                  "line_id": "33333333-3333-4333-8333-333333333333",
                  "line_no": 1,
                  "code": "000123",
                  "name": "演習商品",
                  "quantity": 1,
                  "unit_price": "100",
                  "tax_rate": "0.10",
                  "discount": null,
                  "discount_per_unit": "0",
                  "net_unit_price": "100",
                  "line_subtotal": "100"
                }
              ],
              "subtotal": "100",
              "total": "110",
              "taxes": [
                {
                  "tax_rate": "0.10",
                  "taxable_subtotal": "100",
                  "tax_amount": "10"
                }
              ],
              "amounts_are_reference": false
            },
            "purchase_status": "NOT_REQUESTED",
            "purchase": null,
            "operation_id": "22222222-2222-4222-8222-222222222222",
            "operation_status": "APPLIED",
            "applied_version": "2",
            "code": null
          },
          {
            "cart": {
              "cart_id": "11111111-1111-4111-8111-111111111111",
              "version": "4",
              "state": "SAVED",
              "staff_id": "staff01",
              "member_state": "UNSPECIFIED",
              "member_id": null,
              "pending_member_id": null,
              "lines": [
                {
                  "line_id": "33333333-3333-4333-8333-333333333333",
                  "line_no": 1,
                  "code": "000123",
                  "name": "演習商品",
                  "quantity": 1,
                  "unit_price": "100",
                  "tax_rate": "0.10",
                  "discount": null,
                  "discount_per_unit": "0",
                  "net_unit_price": "100",
                  "line_subtotal": "100"
                }
              ],
              "subtotal": "100",
              "total": "110",
              "taxes": [
                {
                  "tax_rate": "0.10",
                  "taxable_subtotal": "100",
                  "tax_amount": "10"
                }
              ],
              "amounts_are_reference": false
            },
            "purchase_status": "SAVED",
            "purchase": {
              "cart_id": "11111111-1111-4111-8111-111111111111",
              "staff_id": "staff01",
              "member_id": null,
              "purchased_at": "2026-09-27T01:00:00.000000Z",
              "lines": [
                {
                  "line_no": 1,
                  "code": "000123",
                  "name": "演習商品",
                  "quantity": 1,
                  "unit_price": "100",
                  "tax_rate": "0.10",
                  "discount": null,
                  "discount_per_unit": "0",
                  "net_unit_price": "100",
                  "line_subtotal": "100"
                }
              ],
              "subtotal": "100",
              "total": "110",
              "taxes": [
                {
                  "tax_rate": "0.10",
                  "taxable_subtotal": "100",
                  "tax_amount": "10"
                }
              ]
            },
            "operation_id": "22222222-2222-4222-8222-222222222222",
            "operation_status": "APPLIED",
            "applied_version": "4",
            "code": null
          },
          {
            "cart": {
              "cart_id": "11111111-1111-4111-8111-111111111111",
              "version": "4",
              "state": "UNSAVED",
              "staff_id": "staff01",
              "member_state": "UNSPECIFIED",
              "member_id": null,
              "pending_member_id": null,
              "lines": [
                {
                  "line_id": "33333333-3333-4333-8333-333333333333",
                  "line_no": 1,
                  "code": "000123",
                  "name": "演習商品",
                  "quantity": 1,
                  "unit_price": "100",
                  "tax_rate": "0.10",
                  "discount": null,
                  "discount_per_unit": "0",
                  "net_unit_price": "100",
                  "line_subtotal": "100"
                }
              ],
              "subtotal": "100",
              "total": "110",
              "taxes": [
                {
                  "tax_rate": "0.10",
                  "taxable_subtotal": "100",
                  "tax_amount": "10"
                }
              ],
              "amounts_are_reference": false
            },
            "purchase_status": "UNSAVED",
            "purchase": null,
            "operation_id": "22222222-2222-4222-8222-222222222222",
            "operation_status": "APPLIED",
            "applied_version": "4",
            "code": null
          }
        ]
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          },
          "current": {
            "$ref": "#/components/schemas/OperationResult"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "description": "currentは認証・所有関係を確認でき、DB照合結果がある場合だけ返す。401/403では含めない。金額やSQL例外を反射しない。HTTP失敗だけで未保存とは断定しない。",
        "examples": [
          {
            "code": "VERSION_CONFLICT",
            "message": "別の操作が反映されました。最新状態を確認してください。"
          }
        ]
      },
...
    }
  }
}